Integer -> Long

+++ app/src/main/java/com/example/investmentinconstruction/AdapterState/FinalGame/FinalGameState.java
@@ -3,9 +3,9 @@ package com.example.investmentinconstruction.AdapterState.FinalGame;
 public class FinalGameState {
 
     private String name; // ныне uid
-    private Integer fullProfit;
+    private Long fullProfit;
     private int picture;
-    public FinalGameState(String name, Integer fullProfit, int picture) {
+    public FinalGameState(String name, Long fullProfit, int picture) {
         this.name = name;
         this.fullProfit = fullProfit;
         this.picture = picture;
@@ -19,11 +19,11 @@ public class FinalGameState {
         this.name = name;
     }
 
-    public Integer getFullProfit() {
+    public Long getFullProfit() {
         return fullProfit;
     }
 
-    public void setFullProfit(Integer fullProfit) {
+    public void setFullProfit(Long fullProfit) {
         this.fullProfit = fullProfit;
     }
